+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * Ignore backup files in `projectile-get-other-files`.
 * Ignore Ensime cache directory, `.ensime_cache`.
 * [#364](https://github.com/bbatsov/projectile/issues/364): `projectile-add-known-project` can now be used interactively.
+* `projectile-mode` is now a global mode.
 
 ### Bugs fixed